remove `details` from annotation 🖍️ (#7)
Remove the `details` field from all annotations. In certain conditions,
including this field (which was filled with a marshalled JSON string)
appears to be causing a validation error in posting an initial check
creation to GitHub. In the reproduced error cases, the post body
remained valid JSON, but nonetheless removing the `details` field
resolved the validation issue.