Feature: allow to pass the app key through an environment variable
This adds the possibility to supply the GitHub application's private key through a new environment variable: `APP_KEY`. This is useful in CI/CDs where secrets may not be stored in files, and users may not wish to store them into an intermediary file.

Note: there is no validation whether or not a private key was supplied as there wasn't any check in the original code either.