Add non-distroless stage to Dockerfile
The `distroless` flavor is useful to keep image size to the bare minimum but it can be problematic for certain type of work. For example, if the `distroless` flavor is used inside a GitHub Workflow with `jobs.<job_id>.container` (instead of `jobs.<job_id>.steps[*].run` containing `docker run <distroless-image> [...]`) as GitHub will try to run system commands such as `tail` which will fail.

This thus adds a non-distroless flavor that can be used if the user intents to use system binaries (package managers, shells, etc.)